+++ src/observe/_observe.js
@@ -2,6 +2,8 @@
 // exporting modules to be included the UMD bundle
 
 import interactionType from './interaction-type';
+import shadowMutations from './shadow-mutations';
 export default {
   interactionType,
+  shadowMutations,
 };