+++ test/intern.js
@@ -104,6 +104,7 @@ define([
       'test/unit/maintain.hidden.test',
       'test/unit/maintain.tab-focus.test',
       'test/unit/observe.interaction-type.test',
+      'test/unit/observe.shadow-mutations.test',
       'test/unit/prototype.element.prototype.matches.test',
       'test/unit/prototype.svgelement.prototype.focus.test',
       'test/unit/prototype.window.customevent.test',